python-ocr/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
paddleocr>=2.7.0
paddlepaddle>=2.6.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
numpy>=1.26.0
opencv-python>=4.9.0
requests>=2.31.0
python-dotenv>=1.0.0
